Add ability to override patron e-mails.
Valid e-mails are generally replaced after migrating production data into a development environment.
This workflow has the potential for re-adding authentic production e-mails.

This provides a way to forcibly guarantee that a stub/alternative e-mail is used for every single account to be imported.
This adds only a single e-mail address to all accounts being transformed based on a build time variable.