RTR 464-469: Add Example TokenUtility Script.
This provides an example script that utilizes the `TokenUtility` get Access and get Refresh token functions.

This example operates the workflow similar to that of a unit test.
This handles several exceptions and provides a final report.

This also utilizes the `initialContext` to provide a default `logLevel`.
This shows how the `logLevel` can be used for workflows meant to be run directly as opposed to curl requests.