RTR 464-469: Remove extra new line.
The headers generally always have two lines before them to make the plain text easier to read and navigate.

In the specific case of a header followed immediately by a child header, such as:
```
## Parent

### Child
```

The extra new line is not necessary and is avoided in this specific `README.md`.

Remove the extra new line to be consistent with this behavior.