RTR 464-472: Okapi changes.
Refactor `okapi-internal` to `okapi-url` to be more clear and consistent.
There is documentation regarding `okapiUrl` but its only use is a console log in `hegis-po`.
Remove both the reference to the `okapiUrl` and the console log in `hegis-po`.

Rename all Okapi Login JSON files to `folioLogin.json` to make things consistent.
Update descriptions and names.

Switch the `X-Okapi-Token` `headerOutputVariables` processing to `Set-Cookie` with the `asArray` set to `true`.
This is now a `LOCAL` varaible because it should not be exposed to the top-level of the process.
THe  `loginResponse` should also be `LOCAL` for the same reasons.

The **Get Access Token** is a direct copy of the logic from the **Example FOLIO Login** workflow.