RTR 464-473: Add example login workflow.
This logs into FOLIO (via the `okapi-url`).
This extracts the `Set-Cookie` HTTP header.
The header searched and the `folioAccessToken` is extracted if found.
This Access Token gets saved as the `X-Okapi-Token` and is then printed.

Two different scripts are used in order to demonstrate that the `X-Okapi-Token` does properly propagate.