Change prepare holdings Javascript to utilize list of allowed keys.
Provide a new built time variable `holdingsRecordKeys` for providing an list of the allow keys (as a JSON array).

The list of allowed keys is generated from [mod-inventory-storage holdingsRecord.json](https://github.com/folio-org/mod-inventory-storage/blob/master/ramls/holdings-storage/holdingsRecord.json).

Use the `jq` program as follows to extract the array of allowed keys along with the `fw` script to set assign the `holdingsRecordKeys` value:

```shell
fw config set holdingsRecordKeys $(jq -Mc '.properties | keys' holdingsRecord.json)
```

Recent changes to Camunda allow for using more up to date JavaScript functionality.
Update the `prepareHoldings.js` to utilize the newer code.
Provide a `main()` function in case future needs require silently returning without error (allowing to avoid excessive nesting).

Use the newer JavaScript functionality to implement the solution.

This list is implemented as a `Set()`.
This allows for passing the `holdingsRecordKeys` literally into the code even when not defined.
When not defined this would simply initialize an empty `Set()`.

Walk through each key in the retrieved holdings object and if that key is not found then delete it.
Debug logging is provided to log any deletions.